+++ src/js/components/molecule/__tests__/BadgeWithText.test.js
@@ -7,33 +7,28 @@ import BadgeWithText from '../BadgeWithText';
 configure({ adapter: new Adapter() });
 
 describe('BadgeWithText', () => {
-  it('renders default color', () => {
-    const wrapper = shallow(<BadgeWithText number={10}>Default</BadgeWithText>);
-    expect(wrapper).toMatchSnapshot();
+  it('renders with number', () => {
+    const wrapper = shallow(<BadgeWithText badgePrimary number={121} />);
+    const findingText = wrapper.find('Badge');
+    expect(findingText.html())
+      .toBe('<span class="\">121</span>');
   });
 
-  it('renders primary color', () => {
-    const wrapper = shallow(<BadgeWithText badgePrimary number={10}>badgePrimary</BadgeWithText>);
-    expect(wrapper).toMatchSnapshot();
+  it('renders with text', () => {
+    const wrapper = shallow(<BadgeWithText badgePrimary number={121}>Lorem ipsum.</BadgeWithText>);
+    const findingText = wrapper.find('Text');
+    expect(findingText.html())
+      .toBe('<p class=" ">Lorem ipsum.</p>');
   });
 
-  it('renders success color', () => {
-    const wrapper = shallow(<BadgeWithText badgeSuccess number={10}>badgeSuccess</BadgeWithText>);
-    expect(wrapper).toMatchSnapshot();
-  });
-
-  it('renders danger color', () => {
-    const wrapper = shallow(<BadgeWithText badgeDanger number={10}>badgeDanger</BadgeWithText>);
-    expect(wrapper).toMatchSnapshot();
-  });
-
-  it('renders grayLight color', () => {
-    const wrapper = shallow(<BadgeWithText badgeGrayLight number={10}>badgeGrayLight</BadgeWithText>);
-    expect(wrapper).toMatchSnapshot();
-  });
-
-  it('renders with children inside it', () => {
-    const wrapper = shallow(<BadgeWithText number={10}><div>Lorem <strong>Ipsum</strong></div></BadgeWithText>);
-    expect(wrapper.children().html()).toBe('<span><span class="\">10</span><p class=" \"><div>Lorem <strong>Ipsum</strong></div></p></span>');
+  it('renders with children inside it (number/text)', () => {
+    const wrapper = shallow(
+      <BadgeWithText badgeDanger number={10}>
+        <div>Lorem <strong>Ipsum</strong></div>
+      </BadgeWithText>,
+    );
+    expect(wrapper.children()
+      .html())
+      .toBe('<span><span class="\">10</span><p class=" \"><div>Lorem <strong>Ipsum</strong></div></p></span>');
   });
 });